Cleanup error handling, renaming
