+++ src/main.rs
@@ -1,15 +1,15 @@
+mod app;
 mod config;
 mod opt;
 mod registry;
-mod runner;
 mod util;
 
 use clap::Clap;
 use colored::Color::{self, *};
 
+use app::App;
 use config::Config;
 use opt::Opts;
-use runner::CommandRunner;
 
 /// Default max depth passed to [GlobWalker](globwalker::GlobWalker)
 pub const DEFAULT_MAX_DEPTH: usize = 2;
@@ -33,7 +33,7 @@ pub const DEFAULT_COLORS: &[Color] = &[
 fn main() {
     let config = Config::load_default_location().unwrap_or_default();
 
-    if let Err(e) = CommandRunner::run(Opts::parse(), config) {
-        eprintln!("{}", e.to_string());
+    if let Err(e) = App::run(Opts::parse(), config) {
+        eprintln!("{}", e);
     }
 }
